Implements leaderboard database. Creates a new user entry when a user registers their account, and handles repeat usernames/emails.  Updates leaderborad page to show fetch from the database.
Still need to adapt the system to use the userToken that is generated at user registration and stored in the leaderboard database. It still generates a new userToken regardless if you have an account. This will require a decently large overhaul to account for the async nature of fetching a userToken from the database.